requirements.txt

pyzmq==27.1.0
fastapi==0.135.3
pyyaml==6.0.3
uvicorn==0.38.0
tornado>=6.4.0
requests>=2.32.0
python-dotenv>=1.0.1
reportlab>=4.2.2
PyJWT[crypto]>=2.9.0

# Network security modules (Phase 8)
scapy>=2.5.0
